Refactor seekerController and seekerRoutes
