+++ Back-End/controllers/index.js
@@ -1,0 +1,11 @@
+const signUpSeeker = require("./seeker/signUpSeeker");
+const verifySignUpOtp = require("./seeker/verifySignUpOtp");
+const resendOtp = require("./seeker/resendOtp");
+const signInSeeker = require("./seeker/signInSeeker");
+
+module.exports = {
+  signUpSeeker,
+  verifySignUpOtp,
+  resendOtp,
+  signInSeeker,
+};
