+++ Back-End/routes/seekerRoutes.js
@@ -1,9 +1,14 @@
 const router = require("express").Router();
-const SeekerController = require("../controllers/seekerController.js");
+const {
+  signUpSeeker,
+  verifySignUpOtp,
+  resendOtp,
+  signInSeeker,
+} = require("../controllers");
 
-router.post("/signup", SeekerController.signUpSeeker);
-router.post("/verifySignUpOtp", SeekerController.verifySignUpOtp);
-router.post("/signin", SeekerController.signInSeeker);
-router.post("/resendOtp", SeekerController.resendOtp);
+router.post("/signup", signUpSeeker.signUpSeeker);
+router.post("/verifySignUpOtp", verifySignUpOtp.verifySignUpOtp);
+router.post("/resendOtp", resendOtp.resendOtp);
+router.post("/signin", signInSeeker.signInSeeker);
 
 module.exports = router;
